requirements.txt
# ── LLM & Agent Framework ──
langchain>=0.3.0
langchain-ollama>=0.3.0
langgraph>=0.2.0

# ── MCP (Model Context Protocol) ──
mcp>=1.0.0

# ── Google APIs ──
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ── Notion ──
notion-client>=2.2.0

# ── Utilities ──
python-dotenv>=1.0.0
pydantic>=2.0.0

# ── Optional: API server ──
# fastapi>=0.110.0
# uvicorn>=0.30.0
